clean up decimal array module
