+++ vortex-array/src/arrays/decimal/vtable/serde.rs
