+++ vortex-array/src/arrays/decimal/vtable/visitor.rs
@@ -1,0 +1,16 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: Copyright the Vortex contributors
+
+use crate::arrays::{DecimalArray, DecimalVTable};
+use crate::vtable::{ValidityHelper, VisitorVTable};
+use crate::{ArrayBufferVisitor, ArrayChildVisitor};
+
+impl VisitorVTable<DecimalVTable> for DecimalVTable {
+    fn visit_buffers(array: &DecimalArray, visitor: &mut dyn ArrayBufferVisitor) {
+        visitor.visit_buffer(&array.values);
+    }
+
+    fn visit_children(array: &DecimalArray, visitor: &mut dyn ArrayChildVisitor) {
+        visitor.visit_validity(array.validity(), array.len())
+    }
+}
