+++ vortex-array/src/builders/mod.rs
@@ -35,7 +35,7 @@ use vortex_error::{VortexResult, vortex_panic};
 use vortex_mask::Mask;
 use vortex_scalar::{Scalar, match_each_decimal_value_type};
 
-use crate::arrays::smallest_storage_type;
+use crate::arrays::smallest_decimal_value_type;
 use crate::canonical::Canonical;
 use crate::{Array, ArrayRef};
 
@@ -246,7 +246,7 @@ pub fn builder_with_capacity(dtype: &DType, capacity: usize) -> Box<dyn ArrayBui
             })
         }
         DType::Decimal(decimal_type, n) => {
-            match_each_decimal_value_type!(smallest_storage_type(decimal_type), |D| {
+            match_each_decimal_value_type!(smallest_decimal_value_type(decimal_type), |D| {
                 Box::new(DecimalBuilder::with_capacity::<D>(
                     capacity,
                     *decimal_type,
